+++ ecs/world_examples_test.go
@@ -281,7 +281,7 @@ func ExampleWorld_SetListener() {
 	world.SetListener(listener)
 
 	world.NewEntity()
-	// Output: &{{1 0} {[0 0 0 0]} {[0 0 0 0]} [] [] [] 1 {0 0} {0 0} false}
+	// Output: &{{1 0} {[0 0 0 0]} [] [] <nil> <nil> {0 0} 1 false false}
 }
 
 func ExampleWorld_Stats() {